requirements.txt
tensorflow
tflearn
nltk
flask
Flask-WTF
wtforms[email]
flask-sqlalchemy
Flask-PyMongo
gunicorn
alembic==1.4.2
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Gravatar==0.5.0
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.2
marshmallow-sqlalchemy==0.23.0
PyJWT==2.4.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
